Rename `country_registry.rb` to match class name `country_register.rb`
