+++ lib/dfid-transition/patch/countries.rb
@@ -1,14 +1,14 @@
 require 'rest-client'
 require 'json'
 require 'dfid-transition/patch/base'
-require 'govuk/country_registry'
+require 'gov_uk/registers/country'
 
 module DfidTransition
   module Patch
     class Countries < Base
       def mutate_schema
         country_facet['allowed_values'] = transform_to_label_value(
-          Govuk::CountryRegister.countries)
+          GovUk::Registers::Country.countries)
       end
 
     private
